[clean] モジュールに分割 環境変数を使用するよう変更
モジュールに分割しました
チャンネルを環境変数で指定するように変更しました
